app/requirements.txt
flask
boto3
psycopg2-binary
python-dotenv